Move individual hook features to own files
